cli/ingredient-parser/requirements.txt
ingredient-parser-nlp==0.1.0b4
nltk
typer
